refactor: Major refactor to use IDK services part7. JS fixes mostly

+++ modules/openid-federation-common/src/commonMain/kotlin/com/sphereon/openid/fed/common/builder/EntityConfigurationStatementObjectBuilder.kt
@@ -8,8 +8,8 @@ import kotlinx.serialization.json.JsonObject
 
 class EntityConfigurationStatementObjectBuilder {
     private var iss: String? = null
-    private var exp: Int? = null
-    private var iat: Int? = null
+    private var exp: Double? = null
+    private var iat: Double? = null
     private lateinit var jwks: List<Jwk>
     private var metadata: MutableMap<String, JsonObject> = mutableMapOf()
     private var metadataPolicy: MutableMap<String, JsonObject> = mutableMapOf()
@@ -20,8 +20,8 @@ class EntityConfigurationStatementObjectBuilder {
     private val trustMarks: MutableList<TrustMark> = mutableListOf()
 
     fun iss(iss: String) = apply { this.iss = iss }
-    fun exp(exp: Int) = apply { this.exp = exp }
-    fun iat(iat: Int) = apply { this.iat = iat }
+    fun exp(exp: Double) = apply { this.exp = exp }
+    fun iat(iat: Double) = apply { this.iat = iat }
     fun jwks(jwks: List<Jwk>) = apply { this.jwks = jwks }
 
     fun metadata(metadata: Pair<String, JsonObject>) = apply {
